구 템플릿 deprecation (#2220)
템플릿 안녕~
