+++ apps/mobile/lib/screens/editor/editor_query.graphql
@@ -42,15 +42,5 @@ query EditorScreen_Query($slug: String!, $siteId: ID!) {
 
   site(siteId: $siteId) {
     id
-
-    templates {
-      id
-      title
-
-      entity {
-        id
-        slug
-      }
-    }
   }
 }
