+++ apps/website/src/routes/website/(dashboard)/[slug]/Editor.svelte
@@ -49,7 +49,6 @@
   import Limit from './Limit.svelte';
   import PasteModal from './PasteModal.svelte';
   import { YState } from './state.svelte';
-  import TemplateModal from './TemplateModal.svelte';
   import type { Editor } from '@tiptap/core';
   import type { PageLayout, Ref } from '@typie/ui/utils';
   import type { Editor_query } from '$graphql';
@@ -137,7 +136,6 @@
             }
 
             ...Editor_Limit_site
-            ...Editor_Placeholder_site
             ...Editor_TopToolbar_site
           }
 
@@ -1208,7 +1206,6 @@
                           bind:editor
                         />
                         {#if editor && mounted}
-                          <TemplateModal $site={entity.site} {doc} {editor} {focused} />
                           {#if app.preference.current.lineHighlightEnabled}
                             <Highlight {editor} scale={editorScale} />
                           {/if}
