+++ packages/ui/src/tiptap/extensions/placeholder.ts
@@ -1,72 +1,53 @@
 import { Extension } from '@tiptap/core';
 import { Plugin, PluginKey } from '@tiptap/pm/state';
 import { Decoration, DecorationSet } from '@tiptap/pm/view';
-import { css } from '@typie/styled-system/css';
-import { match } from 'ts-pattern';
+import { mount, unmount } from 'svelte';
 import { isBodyEmpty } from '../lib';
+import PlaceholderWidget from './PlaceholderWidget.svelte';
 import type { EditorState } from '@tiptap/pm/state';
 
 export const Placeholder = Extension.create({
   name: 'placeholder',
-  priority: 1000,
 
   addProseMirrorPlugins() {
     const key = new PluginKey('placeholder');
+    const componentInstances = new Map<HTMLElement, ReturnType<typeof mount>>();
 
     const createDecorations = (state: EditorState) => {
-      if (!this.editor.isEditable || !this.editor.isFocused || window.__webview__) {
+      if (!this.editor.isEditable || !isBodyEmpty(state)) {
         return DecorationSet.empty;
       }
 
-      const { selection } = state;
-      const { $anchor, empty } = selection;
-
-      if (!empty) {
+      const firstParagraph = state.doc.child(0)?.child(0);
+      if (!firstParagraph || firstParagraph.type.name !== 'paragraph') {
         return DecorationSet.empty;
       }
 
-      if ($anchor.parent.childCount > 0) {
-        return DecorationSet.empty;
-      }
+      const decoration = Decoration.widget(
+        2,
+        () => {
+          const container = document.createElement('div');
+          container.contentEditable = 'false';
+          container.style.position = 'absolute';
+          container.style.top = 'calc(var(--prosemirror-page-margin-top) + var(--prosemirror-padding-top))';
+          container.style.left = 'calc(var(--prosemirror-page-margin-left) + var(--prosemirror-padding-x))';
+          container.style.maxWidth = 'var(--prosemirror-max-width)';
+          container.style.pointerEvents = 'none';
 
-      let placeholder: string | null = null;
-      let from: number | null = null;
-      let to: number | null = null;
+          const component = mount(PlaceholderWidget, {
+            target: container,
+          });
 
-      if (
-        $anchor.depth === 2 &&
-        $anchor.parent.type.name === 'paragraph' &&
-        ($anchor.parent.attrs.textAlign === 'left' || $anchor.parent.attrs.textAlign === 'justify')
-      ) {
-        if (!isBodyEmpty(state)) {
-          placeholder = window.__webview__ ? '내용을 입력해보세요...' : '내용을 입력하거나 /를 입력해 블록 삽입하기...';
-          from = $anchor.pos <= 2 ? 1 : $anchor.before();
-          to = $anchor.pos <= 2 ? 3 : $anchor.after();
-        }
-      } else if ($anchor.depth >= 2) {
-        const block = $anchor.node(2);
-        if (block) {
-          placeholder = match(block.type.name)
-            .with('blockquote', () => '인용구')
-            .with('callout', () => '강조')
-            .with('fold', () => '접기')
-            .with('bullet_list', 'ordered_list', () => '목록')
-            .otherwise(() => null);
+          componentInstances.set(container, component);
 
-          if (placeholder) {
-            from = $anchor.before();
-            to = $anchor.after();
-          }
-        }
-      }
+          return container;
+        },
+        {
+          side: -1,
+        },
+      );
 
-      if (!placeholder) {
-        return DecorationSet.empty;
-      }
-
-      if (from !== null && to !== null) {
-        return DecorationSet.create(state.doc, [createDecoration(from, to, placeholder)]);
-      }
+      return DecorationSet.create(state.doc, [decoration]);
     };
 
     return [
@@ -76,23 +57,20 @@ export const Placeholder = Extension.create({
           init: (_, state) => {
             return createDecorations(state);
           },
-          apply: (tr, decorations, oldState, newState) => {
-            if (!tr.docChanged && oldState.selection.eq(newState.selection)) {
-              return decorations;
-            }
+          apply: (_, decorations, oldState, newState) => {
+            const oldEmpty = isBodyEmpty(oldState);
+            const newEmpty = isBodyEmpty(newState);
 
-            const oldPos = oldState.selection.$anchor;
-            const newPos = newState.selection.$anchor;
+            if (oldEmpty !== newEmpty) {
+              componentInstances.forEach((component) => {
+                unmount(component);
+              });
+              componentInstances.clear();
 
-            if (
-              oldPos.parent !== newPos.parent ||
-              oldPos.parent.childCount !== newPos.parent.childCount ||
-              !oldState.selection.eq(newState.selection)
-            ) {
               return createDecorations(newState);
             }
 
-            return decorations?.map(tr.mapping, tr.doc);
+            return decorations;
           },
         },
         props: {
@@ -100,24 +78,13 @@ export const Placeholder = Extension.create({
             return this.getState(state);
           },
         },
+        destroy() {
+          componentInstances.forEach((component) => {
+            unmount(component);
+          });
+          componentInstances.clear();
+        },
       }),
     ];
   },
 });
-
-const createDecoration = (from: number, to: number, placeholder: string) => {
-  return Decoration.node(from, to, {
-    class: css({
-      position: 'relative',
-      _before: {
-        content: 'attr(data-placeholder)',
-        position: 'absolute',
-        left: '0',
-        color: 'text.disabled',
-        pointerEvents: 'none',
-        lineClamp: '1',
-      },
-    }),
-    'data-placeholder': placeholder,
-  });
-};
